requirements.txt
# Add project dependencies here.
openai
python-dotenv